third_party/blink/web_tests/platform/generic/plugins/navigator-mimeTypes-length-expected.txt

Test for bug 10038: REGRESSION: Length of navigator.mimeTypes collection returns number of installed plugins, not number of registered mime types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.mimeTypes.length == numberOfMimeTypes is true
PASS navigator.mimeTypes[navigator.mimeTypes.length+1] is undefined
PASS successfullyParsed is true

TEST COMPLETE
